move api call to service, update readme
