+++ automation/get_target_servers.py
@@ -5,33 +5,11 @@ import json
 import argparse
 import gzip
 import os
-from requests import Response
 from requests import Session
-
-from service import apigee_auth
+from service import apigee_auth, apigee_target_servers
 
 # Global session used for all requests.
 REQUEST = Session()
-APIGEE_API_URL = 'https://api.enterprise.apigee.com/v1/organizations/{}/environments/{}'
-
-
-def get_target_server(org_name: str, env_name: str, target_server_name: str):
-    """Retrieves the target server by name of the given organization name and environment name"""
-
-    response = REQUEST.get((APIGEE_API_URL + '/targetservers/{}').format(org_name, env_name, target_server_name))
-
-    if response.status_code == 404:
-        return None
-
-    if response.status_code == 200:
-        return response.json()
-
-    raise Exception(print_error(response))
-
-
-def print_error(response: Response) -> str:
-    """Prints the error returned from an API call"""
-    return f'Error: {response.status_code} - {response.reason}. \n {response.text}'
 
 
 def parse_args():
@@ -122,7 +100,7 @@ def main():
     result = []
 
     for env in environments:
-        target_server = get_target_server(organization, env, target_server_name)
+        target_server = apigee_target_servers.get_target_server(REQUEST, organization, env, target_server_name)
         if target_server is not None:
             result.append({'environment': env, 'target_server': target_server})
